# koan/openapi.yaml
...
components:
  schemas:
    Error:
      properties:
        error:
          properties:
            code:
              type: string
            message:
              type: string
          required:
          - code
          - message
          type: object
      required:
      - error
      type: object
  securitySchemes:
    bearerAuth:
      scheme: bearer
      type: http
info:
  description: Optional, token-authenticated HTTP control layer for the Kōan agent. See docs/operations/rest-api.md.
  title: Kōan REST API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /v1/missions:
    get:
      operationId: missions_list_missions_route_get
      parameters:
      - description: Restrict results to one mission status.
        in: query
        name: status
        required: false
        schema:
          enum:
          - pending
          - in_progress
          - done
          - failed
          - removed
          type: string
      - description: Restrict results to one project.
        in: query
        name: project
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Successful response
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Missing or malformed bearer token
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Invalid bearer token
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error response
      summary: List missions, newest first, optionally filtered.
      tags:
      - missions
    post:
      operationId: missions_create_mission_post
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - properties:
                  command:
                    pattern: \S
                required:
                - command
              - properties:
                  text:
                    pattern: \S
                required:
                - text
              properties:
                command:
                  description: Slash-command mission; takes precedence over text.
                  type: string
                project:
                  description: Optional project name added as a project tag.
                  type: string
                text:
                  description: Free-form mission text.
                  type: string
                urgent:
                  default: false
                  description: Insert the mission at the front of the pending queue.
                  type: boolean
              type: object
        required: true
      responses:
        '202':
          description: Successful response
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Missing or malformed bearer token
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Invalid bearer token
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error response
      summary: Queue a new mission.
      tags:
      - missions
  /v1/missions/reorder:
    post:
      operationId: missions_reorder_mission_route_post
      requestBody:
        content:
          application/json:
            schema:
              properties:
                mission_id:
                  pattern: \S
                  type: string
                target_position:
                  description: One-indexed position in the pending queue.
                  minimum: 1
                  type: integer
              required:
              - mission_id
              - target_position
              type: object
        required: true
      responses:
        '200':
          description: Successful response
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Missing or malformed bearer token
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Invalid bearer token
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error response
      summary: Move a pending mission to a new position.
      tags:
      - missions
  /v1/missions/{mission_id}:
    delete:
      operationId: missions_delete_mission_delete
      parameters:
      - in: path
        name: mission_id
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successful response
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Missing or malformed bearer token
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Invalid bearer token
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error response
      summary: Remove a pending mission.
      tags:
      - missions
    get:
      operationId: missions_get_mission_route_get
      parameters:
      - in: path
        name: mission_id
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successful response
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Missing or malformed bearer token
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Invalid bearer token
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error response
      summary: Fetch one mission by id.
      tags:
      - missions
    patch:
      operationId: missions_edit_mission_patch
      parameters:
      - in: path
        name: mission_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              properties:
                text:
                  pattern: \S
                  type: string
              required:
              - text
              type: object
        required: true
      responses:
        '200':
          description: Successful response
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Missing or malformed bearer token
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Invalid bearer token
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error response
      summary: Change a pending mission.
      tags:
      - missions
  /v1/missions/{mission_id}/result:
    get:
      operationId: missions_get_mission_result_route_get
      parameters:
      - in: path
        name: mission_id
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successful response
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Missing or malformed bearer token
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Invalid bearer token
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error response
      summary: Fetch a finished mission's result.
      tags:
      - missions
...
  /v1/status:
    get:
      operationId: status_status_get
      responses:
        '200':
          description: Successful response
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Missing or malformed bearer token
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Invalid bearer token
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error response
      summary: Get current agent state, execution and mission counters.
      tags:
      - status
...
security:
- bearerAuth: []
servers:
- description: Default loopback bind
  url: http://127.0.0.1:8420
...